refactor: 쿠키로부터 토큰 가져오는 기능 TokenProvider내부로 이동
