+++ src/main/java/roomescape/service/AuthService.java
@@ -1,10 +1,11 @@
 package roomescape.service;
 
+import jakarta.servlet.http.Cookie;
 import org.springframework.stereotype.Service;
-import roomescape.controller.TokenProvider;
 import roomescape.dao.MemberDao;
 import roomescape.domain.Member;
 import roomescape.exception.NotFoundException;
+import roomescape.service.dto.LoginMember;
 import roomescape.service.dto.request.LoginRequest;
 
 @Service
@@ -27,9 +28,12 @@ public class AuthService {
         return tokenProvider.createToken(member);
     }
 
-    public Member findLoginMember(String token) {
+    public LoginMember findLoginMember(Cookie[] cookies) {
+        String token = tokenProvider.extractTokenBy(cookies);
+
         String payload = tokenProvider.getPayload(token);
-        return memberDao.findById(Long.valueOf(payload))
+        Member member = memberDao.findById(Long.valueOf(payload))
                 .orElseThrow(() -> new NotFoundException("사용자를 찾을 수 없습니다."));
+        return new LoginMember(member.getId(), member.getName(), member.getRole());
     }
 }
