+++ src/main/java/roomescape/service/TokenProvider.java
@@ -1,10 +1,14 @@
-package roomescape.controller;
+package roomescape.service;
+
+import static roomescape.controller.AuthController.TOKEN_NAME;
 
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
+import jakarta.servlet.http.Cookie;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import roomescape.domain.Member;
+import roomescape.exception.BadRequestException;
 
 @Component
 public class TokenProvider {
@@ -25,5 +29,15 @@ public class TokenProvider {
                 .build()
                 .parseClaimsJws(token)
                 .getBody().getSubject();
+    }
+
+    public String extractTokenBy(Cookie[] cookies) {
+        for (Cookie cookie : cookies) {
+            if (cookie.getName().equals(TOKEN_NAME)) {
+                return cookie.getValue();
+            }
+        }
+
+        throw new BadRequestException("올바르지 않은 토큰값입니다.");
     }
 }
