+++ src/main/java/roomescape/service/dto/LoginMember.java
@@ -1,13 +1,17 @@
 package roomescape.service.dto;
 
+import roomescape.domain.Role;
+
 public class LoginMember {
 
     private final Long id;
     private final String name;
+    private final Role role;
 
-    public LoginMember(Long id, String name) {
+    public LoginMember(Long id, String name, Role role) {
         this.id = id;
         this.name = name;
+        this.role = role;
     }
 
     public Long getId() {
@@ -16,5 +20,9 @@ public class LoginMember {
 
     public String getName() {
         return name;
+    }
+
+    public Role getRole() {
+        return role;
     }
 }
